replaced os.path with pathlib. put aux metadata functions in a separate module (utils)
