+++ pages/01_metadata.py
@@ -1,11 +1,9 @@
 # Define the page's content within a variable called layout or
 # a function called layout that returns the content.
-#
-# File upload component:
-# https://dash.plotly.com/dash-core-components/upload
+
 
 import base64
-import os
+import pathlib as pl 
 
 import dash
 import dash_bootstrap_components as dbc
@@ -13,86 +11,11 @@ import pandas as pd
 import yaml
 from dash import Input, Output, State, callback, dcc, html
 
+import wazp.utils
+
 ###############
 # Register this page
 dash.register_page(__name__)
-
-
-####################
-# Read yamls in dir as pandas dataframe
-
-
-def df_from_metadata_yaml_files(parent_dir):
-    # TODO: refactor, I think it could be more compact (see below)
-
-    list_metadata_files = [
-        el for el in os.listdir(parent_dir) if el.endswith("metadata.yaml")
-    ]
-
-    list_df_metadata = []
-    for yl in list_metadata_files:
-        with open(os.path.join(parent_dir, yl)) as f:
-            list_df_metadata.append(
-                pd.DataFrame.from_dict(yaml.safe_load(f), orient="index")
-            )
-    return pd.concat(
-        list_df_metadata, ignore_index=True, axis=1
-    ).T  # TODO review why transpose required
-
-
-# TODO: previous approach with json, can I do it as compact?
-# read_fn = lambda x: pd.read_json(os.path.join(parent_dir, x), orient="index")
-# df = pd.concat(map(read_fn, list_metadata_files), ignore_index=True, axis=1)
-
-
-###############
-# Define table component from pandas dataframe
-# TODO: check table style options (see Table class)
-
-
-def generate_tbl_component(parent_dir):
-    # ##########################
-    # Using html dash components
-    # dataframe = df_from_metadata_yaml_files(parent_dir)
-    # table = html.Table(
-    #     children =
-    #     [ # define table head?
-    #        html.Thead(html.Tr([html.Th(col, style={'width':'15%'})
-    #                               for col in dataframe.columns])),
-    #        # define table body
-    #        html.Tbody(
-    #            [
-    #               html.Tr(
-    #                 [
-    #                   html.Td(dataframe.iloc[i][col], style={'width':'15%'})
-    #                   for col in dataframe.columns
-    #                  ]
-    #                 )
-    #                 #range(min(len(dataframe), max_rows))
-    #                 for i in range(len(dataframe))
-    #             ]
-    #         ),
-    #     ],
-    #     style={'width':'100%',}
-    # )
-    #########
-
-    #########
-    # Using bootstrap components
-    table = dbc.Table.from_dataframe(
-        df_from_metadata_yaml_files(parent_dir),
-        bordered=True,
-        hover=False,  # True
-        responsive=True,  # if True, table can be scrolled horizontally?
-        striped=True,  # applies zebra striping to the rows
-        size="sm",
-        style={
-            "width": "100%",
-        },
-    )
-    ############
-
-    return dbc.Container(table, className="p-5")
 
 
 ######################
@@ -132,8 +55,6 @@ layout = html.Div(
 
 #####################################
 # Callbacks
-
-
 @callback(
     Output("output-data-upload", "children"),
     Input("upload-data", "contents"),
@@ -151,6 +72,6 @@ def update_file_drop_output(up_content, up_filename):
             print(e)
             return html.Div(["There was an error processing this file."])
 
-        return generate_tbl_component(
-            video_dir
+        return wazp.utils.generate_tbl_component_from_df(
+            wazp.utils.df_from_metadata_yaml_files(video_dir)
         )  # returns children of 'output-data-upload'
